requirements.txt
aioredis==1.3.1
aniso8601==8.0.0
aredis==1.1.8
astroid==2.4.1
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.2
blinker==1.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
ecdsa==0.14.1
environs==8.0.0
Flask==1.1.2
Flask-Cors==3.0.8
flask-restplus==0.13.0
Flask-SSE==0.2.1
gevent==20.6.2
greenlet==0.4.16
gunicorn==19.9.0
hiredis==1.1.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.7.1
mccabe==0.6.1
passlib==1.7.2
patsy==0.5.1
pyasn1==0.4.8
pycodestyle==2.6.0
pylint==2.5.2
pymongo==3.10.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-jose==3.2.0
pytz==2020.1
redis==3.5.3
rethinkdb==2.4.7
rsa==4.6
Rx==3.1.1
six==1.15.0
threadpoolctl==2.1.0
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
zope.event==4.4
zope.interface==5.1.0
